wrap leadv funcs
